requirements.txt
asgiref==3.2.5
coverage==5.0.4
Django==3.0.4
django-grappelli==2.14.1
django-photologue==3.11
django-sortedm2m==3.0.0
ExifRead==2.1.2
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
mysqlclient=1.4.6

